+++ gitman/commands.py
@@ -6,7 +6,7 @@ import log
 
 from . import common, system
 from .decorators import preserve_cwd
-from .models import Config, Source, load_config
+from .models import Config, Source, find_nested_configs, load_config
 
 
 def init(*, force: bool = False):
@@ -85,10 +85,14 @@ def install(
         ", ".join(names) if names else "<all>",
     )
     count = None
-    config = load_config(root)
 
-    if config:
+    config = load_config(root)
+    configs = [config] if config else []
+    configs.extend(find_nested_configs(root, depth, []))
+    if configs:
         common.newline()
+
+    for config in configs:
         common.show("Installing dependencies...", color="message", log=False)
         common.newline()
         count = config.install_dependencies(
@@ -150,10 +154,14 @@ def update(
         ", ".join(names) if names else "<all>",
     )
     count = None
-    config = load_config(root)
 
-    if config:
+    config = load_config(root)
+    configs = [config] if config else []
+    configs.extend(find_nested_configs(root, depth, []))
+    if configs:
         common.newline()
+
+    for config in configs:
         common.show("Updating dependencies...", color="message", log=False)
         common.newline()
         count = config.install_dependencies(
@@ -223,7 +231,7 @@ def display(*, root=None, depth=None, allow_dirty=True):
     count = None
     config = load_config(root)
 
-    if config:
+    if config and config.root:
         common.newline()
         common.show(
             "Displaying current dependency versions...", color="message", log=False
@@ -231,9 +239,26 @@ def display(*, root=None, depth=None, allow_dirty=True):
         common.newline()
         config.log(datetime.datetime.now().strftime("%F %T"))
         count = 0
+
+        skip_paths = []
         for identity in config.get_dependencies(depth=depth, allow_dirty=allow_dirty):
             count += 1
             config.log("{}: {} @ {}", *identity)
+            skip_paths.append(identity.path)
+
+        nested_configs = find_nested_configs(config.root, depth, skip_paths)
+        if nested_configs:
+            common.show(
+                "Displaying nested dependency versions...", color="message", log=False
+            )
+            common.newline()
+            for nested_config in nested_configs:
+                for identity in nested_config.get_dependencies(
+                    depth=depth, allow_dirty=allow_dirty
+                ):
+                    count += 1
+                    config.log("{}: {} @ {}", *identity)
+
         config.log()
 
     return _display_result("display", "Displayed", count)
@@ -251,10 +276,14 @@ def lock(*names, root=None):
     """
     log.info("Locking dependencies...")
     count = None
-    config = load_config(root)
 
-    if config:
+    config = load_config(root)
+    configs = [config] if config else []
+    configs.extend(find_nested_configs(root, None, []))
+    if configs:
         common.newline()
+
+    for config in configs:
         common.show("Locking dependencies...", color="message", log=False)
         common.newline()
         count = config.lock_dependencies(*names, obey_existing=False)