+++ gitman/models/__init__.py
@@ -1,3 +1,3 @@
 from .source import Source  # isort:skip
-from .config import Config, load_config
+from .config import Config, find_nested_configs, load_config
 from .group import Group